requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-cloud-bigquery==3.14.1
google-cloud-aiplatform==1.38.1
google-generativeai==0.3.2
google-cloud-dataplex==1.10.0
pandas==2.1.4
db-dtypes==1.2.0
streamlit==1.29.0
adk==0.1.0
requests==2.31.0
plotly==5.18.0
openpyxl==3.1.2
xlsxwriter==3.1.9
python-multipart==0.0.6
sqlparse==0.4.4
pyyaml==6.0.1
pytest==7.4.3
pytest-mock==3.12.0
